Add missing std. Add missing cast from int to double.
git-svn-id: https://svn.dealii.org/trunk@8486 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/deal.II/boundaries.cc
@@ -45,7 +45,7 @@ class MySquareFunction : public Function<dim>
     
     virtual double value (const Point<dim>   &p,
 			  const unsigned int  component) const
-      {	return 100*(component+1)*p.square()*sin(p.square()); };
+      {	return 100*(component+1)*p.square()*std::sin(p.square()); };
     
     virtual void   vector_value (const Point<dim>   &p,
 				 Vector<double>     &values) const
@@ -90,7 +90,7 @@ check ()
       GridGenerator::hyper_ball(tr, Point<dim>(), 1);
     }
   else
-    GridGenerator::hyper_cube(tr, -1./sqrt(dim),1./sqrt(dim));
+    GridGenerator::hyper_cube(tr, -1./std::sqrt(static_cast<double>(dim)),1./std::sqrt(static_cast<double>(dim)));
   if (dim != 1)
     {
       static const HyperBallBoundary<dim> boundary;
